requirements.txt
pygame>=2.5.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.4.0
hypothesis>=6.0.0
